requirements.txt
torch
torchvision
scikit-learn
pandas
numpy
pillow
matplotlib
seaborn
streamlit
tqdm
